Make map embedding work from new location
